+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 **New Features:**
 * **[edison-mongo]** Add @ConditionalOnMissingBean to MongoTogglzRepository
 * **[edison-mongo]** Make ID and ETAG constants public to be able to access them from outside
+* **[edison-core]** Allow LDAP authentication for user-defined paths.
 * **[edison-cache]** It is now possible to configure `CaffeineCache` instances as Spring beans and
 use them using `@Cacheable`. For example, this way it is possible, to configure loading caches.
 The new `example-cache` contains a showcase for this. This feature makes the 'old' `CacheRegistry`